app/requirements.txt
python-telegram-bot==21.3
openai>=1.30.0
playwright>=1.44.0
psycopg2-binary>=2.9.9
redis>=5.0.0
boto3>=1.34.0
python-dotenv>=1.0.0
